backend/requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Login==0.6.2
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
numpy==1.26.0
packaging==23.1
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
webargs==8.3.0
Werkzeug==2.3.7
zipp==3.17.0
